backend/requirements.txt
absl-py==2.1.0
aiohappyeyeballs==2.4.4
aiohttp==3.10.5
aiosignal==1.3.1
alabaster==1.0.0
annotated-types==0.7.0
anyio==4.7.0
anytree==2.12.1
apparmor==4.0.3
appdirs==1.4.4
arandr==0.1.11
argcomplete==3.4.0
asciidoc==10.2.1
asgiref==3.8.1
astroid==3.3.8
asttokens==3.0.0
async-timeout==4.0.3
attrs==23.2.1.dev0
autocommand==2.2.2
Automat==22.10.0
autotiling==1.9.3
Babel==2.15.0
bcrypt==4.2.1
Beaker==1.13.0
beautifulsoup4==4.12.3
beniget==0.4.2.post1
black==24.10.0
bleach==6.2.0
blinker==1.8.2
booleanOperations==0.9.0
Brlapi==0.8.6
Brotli==1.1.0
btrfsutil==6.12
build==1.2.2
CacheControl==0.14.1
cachetools==5.5.0
cachy==0.3.0
Cerberus==1.3.5
certifi==2024.12.14
cffi==1.17.1
chardet==5.2.0
charset-normalizer==3.4.1
cheroot==10.0.1
CherryPy==18.10.0
cleo==2.1.0
click==8.1.7
cmd2==2.4.3
colorama==0.4.6
comm==0.2.2
commonmark==0.9.1
ConfigArgParse==1.5.5
configobj==5.0.9
constantly==23.10.4
contourpy==1.3.1
coverage==7.6.9
crashtest==0.4.1
cryptography==44.0.0
CT3==3.3.3
cupshelpers==1.0
cycler==0.12.1
cymem==2.0.8
cython_lint==0.16.0
daemonize==2.5.0
dasbus==1.7
dbus-python==1.3.2
decorator==5.1.1
defcon==0.10.3
defusedxml==0.7.1
Deprecated==1.2.15
dill==0.3.9
dirty-equals==0.8.0
distlib==0.3.9
distro==1.9.0
Django==5.1.4
django-environ==0.11.2
django-extensions==3.2.3
dnf==4.21.1
dnspython==2.6.1
docopt==0.6.2
docutils==0.21.2
dulwich==0.22.7
editables==0.5
entrypoints==0.4
et_xmlfile==2.0.0
evdev==1.7.1
executing==2.1.1.dev0+g8827340.d20241128
fabric==3.2.2
fastapi==0.115.6
fastjsonschema==2.21.1
filelock==3.16.1
fissix==24.4.24
Flask==3.1.0
flit_core==3.10.1
fluidity-sm==0.2.1
fontMath==0.9.4
fontParts==0.12.3
fontPens==0.2.4
fonttools==4.55.3
frozenlist==1.4.1
fs==2.4.16
FVS==0.3.4
gast==0.6.0
google-api-core==2.19.1
google-api-python-client==2.140.0
google-auth==2.36.1
google-auth-httplib2==0.2.0
google-auth-oauthlib==1.2.1
googleapis-common-protos==1.63.2
gpg==1.24.1
graphviz==0.20.3
greenlet==3.1.1
grpcio==1.67.1
grpcio-tools==1.67.1
hatchling==1.25.0
hid==1.0.6
html5lib==1.1
httplib2==0.22.0
hyperlink==21.0.0
hypothesis==6.122.6
i3ipc==2.2.1
icoextract==0.1.5
idna==3.10
ifaddr==0.2.0
imagesize==1.4.1
importlib_metadata==7.2.1
incremental==22.10.0
inflect==7.5.0
iniconfig==2.0.0
installer==0.7.0
invoke==2.2.0
ipykernel==6.29.5
ipython==8.31.0
isodate==0.7.2
isort==5.13.2
itsdangerous==2.1.2
jaraco.classes==3.4.0
jaraco.collections==5.0.1
jaraco.context==5.3.0
jaraco.functools==4.1.0
jaraco.text==4.0.0
jedi==0.19.2
jeepney==0.8.0
Jinja2==3.1.5
joblib==1.4.2
jsonpatch==1.33
jsonpointer==3.0.0
jsonschema==4.23.0
jsonschema-specifications==2024.10.1
jupyter_client==8.6.3
jupyter_core==5.7.2
jupyterlab_pygments==0.3.0
keyring==25.5.0
keyutils==0.6
kiwisolver==1.4.5
lark==1.2.2
lazy-object-proxy==1.10.0
legacy-cgi==2.6.1
lensfun==0.3.4
Levenshtein==0.26.1
lexicon==2.0.1
LibAppArmor==4.0.3
libcomps==0.1.21
libdnf==0.73.4
libfdt==1.7.2
libtorrent==2.0.10
lit==18.1.8.dev0
llvmlite==0.44.0rc1
lockfile==0.12.2
logutils==0.3.5
louis==3.32.0
lxml==5.3.0
Mako==1.3.8.dev0
Markdown==3.7
markdown-it-py==3.0.0
MarkupSafe==2.1.5
matplotlib==3.9.3
matplotlib-inline==0.1.7
mccabe==0.7.0
mdurl==0.1.2
mercurial==6.9
meson==1.6.1
meson-python==0.17.1
mistune==3.1.0
moddb==0.11.0
more-itertools==10.5.0
moreorless==0.4.0
msgpack==1.0.5
multidict==6.0.5
MutatorMath==3.0.1
mypy_extensions==1.0.0
natsort==8.4.0
nbclient==0.10.2
nbconvert==7.16.4
nbformat==5.10.4
nest_asyncio==1.6.0
netaddr==1.3.0
netifaces==0.11.0
netsnmp-python==1.0a1
nftables==0.1
nh3==0.2.19
nose==1.3.7
notify2==0.3.1
nspektr==0.5.0
numba==0.61.0rc1+0.g4fd17fa6d.dirty
numpy==2.2.1
oauth2client==4.1.3
oauthlib==3.2.2
openpyxl==3.1.5
optimus-manager==1.5
ordered-set==4.1.0
orjson==3.10.13
packaging==24.2
pandas==2.2.3
pandocfilters==1.5.1
paramiko==3.5.0
parso==0.8.4
pathspec==0.12.1
pathvalidate==3.2.1
patool==3.1.0
pecan==1.5.1
pefile==2024.8.26
pendulum==3.0.0
perf==0.1
pexpect==4.9.0
phonenumbers==8.13.52
pickleshare==0.7.5
pillow==11.0.0
pipenv==2024.4.0
pipx==1.7.1
pkginfo==1.12.0
platformdirs==4.3.6
pluggy==1.5.0
ply==3.11
poetry==1.8.5
poetry-core==1.9.1
poetry-plugin-export==1.8.0
pooch==1.8.2
portend==3.2.0
prettytable==3.11.0
prompt_toolkit==3.0.48
protobuf==5.29.2
psutil==6.1.1
ptyprocess==0.7.0
pulsemixer==1.5.1
pure_eval==0.2.3
pwquality==1.4.5
pyasn1==0.6.0
pyasn1_modules==0.4.0
pyasyncore==1.0.4
pybind11==2.13.6
pycairo==1.27.0
pyclipper==1.3.0.post6
pycodestyle==2.12.1
pycountry==24.6.1
pycparser==2.22
pycryptodomex==3.21.0
pycups==2.0.4
pycurl==7.45.3
pydantic==2.10.4
pydantic-extra-types==2.10.1
pydantic-settings==2.7.0
pydantic_core==2.27.2
pydot==3.0.1
pyflakes==3.2.0
pygame==2.6.1
Pygments==2.18.0
PyGObject==3.50.0
pyinotify==0.9.6
PyJWT==2.9.0
pylint==3.3.3
pymongo==4.10.1
PyNaCl==1.5.0
PyOpenGL==3.1.7
pyOpenSSL==24.3.0
pyparsing==3.1.2
pyparted==3.13.0
pyperclip==1.9.0
pyproj==3.7.0
pyproject-metadata==0.9.0
pyproject_hooks==1.2.0
PyQt5==5.15.11
PyQt5_sip==12.16.1
PyQt6==6.8.0
PyQt6-3D==6.8.0
PyQt6-Charts==6.8.0
PyQt6-DataVisualization==6.8.0
PyQt6-NetworkAuth==6.8.0
PyQt6-QScintilla==2.14.1
PyQt6-WebEngine==6.8.0
PyQt6_sip==13.9.1
pyroute2==0.7.12
PySide6==6.8.1.1
PySocks==1.7.1
pyte==0.8.2
pytest==8.3.4
pytest-cov==6.0.0
pytest-timeout==2.3.1
python-dateutil==2.9.0
python-dotenv==1.0.1
python-ulid==3.0.0
python-xlib==0.33
pythran==0.17.0
pytools==2024.1.5
pytz==2024.2
pyudev==0.24.3
pywal==3.3.0
pyxdg==0.28
PyYAML==6.0.2
pyzmq==26.2.0
QtPy==2.4.2
RapidFuzz==3.11.0
rdflib==7.1.1
readme_renderer==44.0
referencing==0.35.1
Reflector==2023.6.28.0.36.1
regex==2024.11.6
requests==2.32.3
requests-futures==1.0.1
requests-oauthlib==1.3.1
requests-toolbelt==1.0.0
requests-unixsocket==0.3.0
resolvelib==1.1.0
retrying==1.3.4
rfc3986==2.0.0
rich==13.9.4
rpds-py==0.22.3
rpm==4.20.0
rsa==4.9
ruamel.yaml==0.18.6
ruamel.yaml.clib==0.2.12
scapy==2.6.1
scikit-build==0.17.6
scikit-learn==1.6.0
scipy==1.14.1
scour==0.38.2
SecretStorage==3.3.3
semantic_version==2.10.0
setproctitle==1.3.4
setuptools==75.2.0
setuptools-rust==1.10.2
setuptools-scm==8.1.0
shellingham==1.5.4
shiboken6==6.8.1.1
shiboken6-generator==6.8.1.1
simplejson==3.19.2
six==1.16.0
smbus==1.1
sniffio==1.3.1
snowballstemmer==2.2.0
solaar==1.1.13
sortedcontainers==2.4.0
soupsieve==2.6
Sphinx==8.1.3
sphinxcontrib-applehelp==2.0.0
sphinxcontrib-devhelp==2.0.0
sphinxcontrib-htmlhelp==2.1.0
sphinxcontrib-jsmath==1.0.1
sphinxcontrib-qthelp==2.0.0
sphinxcontrib-serializinghtml==2.0.0
SQLAlchemy==2.0.36.dev0
sqlparse==0.5.1
stack_data==0.6.3
starlette==0.45.1
stem==1.8.3
tabulate==0.9.0
TBB==0.2
tempora==5.7.0
tenacity==9.0.0
termcolor==2.5.0
testpath==0.6.0
textual==1.0.0
thefuck==3.32
threadpoolctl==3.5.0
thrift==0.20.0
time-machine==2.16.0
tinycss2==1.4.0
tokenize_rt==5.2.0
toml==0.10.2
tomli==2.0.1
tomlkit==0.13.2
tornado==6.4.2
tqdm==4.67.1
traitlets==5.14.3
trove-classifiers==2024.10.21.16
twine==6.0.1
Twisted==24.3.0
typeguard==4.4.1
typer==0.14.0
typing_extensions==4.12.2
tzdata==2024.2
uc-micro-py==1.0.3
udiskie==2.5.7
ufoProcessor==1.13.3
ufw==0.36.2
unicodedata2==15.1.0
Unidecode==1.3.8
uritemplate==4.1.1
urllib3==1.26.20
urwid==2.6.15
userpath==1.9.2
validate==5.0.9
validate-pyproject==0.22
virtualenv==20.28.0
virtualenv-clone==0.5.7
vkbasalt-cli==3.1.1
volatile==2.1.0
waitress==3.0.0
watchdog==3.0.0
wcwidth==0.2.13
webencodings==0.5.1
WebOb==1.8.7
websocket-client==1.8.0
websockets==12.0
WebTest==3.0.0
Werkzeug==3.1.3
wheel==0.45.0
wrapt==1.16.0
wxPython==4.2.2
yarl==1.9.4
yaspin==3.1.0
yt-dlp==2024.12.23
zc.lockfile==3.0.post1
zeroconf==0.132.2
zipp==3.21.0
zope.interface==7.2
zopfli==0.2.3.post1
zstandard==0.23.0
